Made changes to allow pipeline to show a build's display name instead of the number if provided.
Added logic to PipelineBuild.getPipelineVersion to use the build's display name
first and then fail back to number if not provided.
This allows users to set the display name to appropriate value
(i.e. the svn revision number) using the build-name-setter-plugin or alike
and have that value displayed in the pipeline view instead of the build number.

Also updated bpp.jelly to use build.displayName if available before using
build.number.