Adjust the UI to fix a stretching-out issue.
Modify the bpp.jelly file which generates the build pipeline view,
in order to fix an issue which causes the build pipeline view to be
stretched out. When the build parameters field was larger than usual,
the cell would expand and stretch the table containing the build
pipeline. Modify main.css to ensure that the arrow images always
point to the next downstream job.