JENKINS-30801 limit Action objects in re-trigger to CauseAction
Passing all Action objects from previous builds results in unexpected
behavior: certain publisher and Builder rely on them to publish
information about the build, and should not be copied to the next build.

The only reason to copy Action objects is to preserve the cause of the build
in order to keep it in the pipeline and replay how the build was triggered.