JENKINS-38670 update docs, add support for JS in handlebars templates
This is a bit anathema to using handlebars but we're more or less stuck
with it: provide a way for a BuildCardExtention to contribute JS to the
build card, allowing for richer build cards.

Also switched an appendTo() to an append() which during testing did
an evaluation of the JS but only when it was in progress - this
would result in build cards being hidden if the build is in progress.

jQuery docs say these are equivalent, but they are liars.